% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{print.iasi_quarto_project}
\alias{print.iasi_quarto_project}
\title{Print an IASI Quarto project}
\usage{
\method{print}{iasi_quarto_project}(x, ...)
}
\arguments{
\item{x}{An \code{iasi_quarto_project}.}

\item{...}{Additional arguments, currently unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print an IASI Quarto project
}
