% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication.R
\name{print.iasi_quarto_publication}
\alias{print.iasi_quarto_publication}
\title{Print an IASI Quarto publication}
\usage{
\method{print}{iasi_quarto_publication}(x, ...)
}
\arguments{
\item{x}{An \code{iasi_quarto_publication}.}

\item{...}{Additional arguments, currently unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print an IASI Quarto publication
}
