% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{release}
\alias{release}
\title{Collect local IASI release artifacts}
\usage{
release(source = NULL, path = ".", force = FALSE)
}
\arguments{
\item{source}{Optional release source.}

\item{path}{Directory from which IASI projects are discovered.}

\item{force}{Reserved for API symmetry.}
}
\value{
Numeric status code. \code{0} means success.
}
\description{
Collect local IASI release artifacts
}
