% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build}
\alias{build}
\title{Build Quarto publications}
\usage{
build(book = NULL, format = NULL)
}
\arguments{
\item{book}{Book or books to build.}

\item{format}{Output format or formats.}
}
\value{
A build execution plan.
}
\description{
Builds one or more Quarto publications.
}
\details{
\code{build()} automatically adapts to the current working directory.

When executed inside a Quarto project, the current project is built.
When executed from the root of a multibook workspace, one or more books
can be selected.
}
