% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{validate}
\alias{validate}
\title{Validate IASI projects}
\usage{
validate(path = ".")
}
\arguments{
\item{path}{Directory from which IASI projects are discovered.}
}
\value{
Numeric status bitmask.
}
\description{
Validate IASI projects
}
